move asides to their own files
