+++ app/components/CartAside/CartAside.module.css
@@ -1,0 +1,3 @@
+.CartAside {
+  display: none;
+}
