+++ app/components/SearchAside/SearchAside.module.css
@@ -1,0 +1,3 @@
+.SearchAside {
+  display: none;
+}
