+++ app/components/index.ts
@@ -1,6 +1,6 @@
 // Export all components from their respective folders
-export {Header} from './Header';
-export type {HeaderProps} from './Header';
+export {Header} from './Header/Header';
+export type {HeaderProps} from './Header/Header';
 
 export {Footer} from './Footer';
 export type {FooterProps} from './Footer';
